+++ tools/cli/lib/src/utils/project_prerequisites.dart
@@ -266,7 +266,7 @@ class ProjectPrerequisites {
     }
 
     final content = await File(destination).readAsString();
-    if (!content.contains('--delete-conflicting-outputs')) {
+    if (content.contains('--delete-conflicting-outputs')) {
       await _writer.copyFile(
         source: source,
         destination: destination,